Refactor to lib and cli app
