+++ AnyReg.sln
@@ -2,20 +2,40 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.13.35913.81
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AnyReg", "AnyReg\AnyReg.vcxproj", "{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AnyRegCore", "AnyRegCore\AnyRegCore.vcxproj", "{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "AnyRegCli", "AnyRegCli\AnyRegCli.vcxproj", "{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
 		Debug|x64 = Debug|x64
 		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
 		Release|x64 = Release|x64
 		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|Any CPU.Build.0 = Debug|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|x64.ActiveCfg = Debug|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|x64.Build.0 = Debug|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|x86.ActiveCfg = Debug|Win32
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Debug|x86.Build.0 = Debug|Win32
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|Any CPU.ActiveCfg = Release|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|Any CPU.Build.0 = Release|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|x64.ActiveCfg = Release|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|x64.Build.0 = Release|x64
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|x86.ActiveCfg = Release|Win32
+		{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}.Release|x86.Build.0 = Release|Win32
+		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|Any CPU.ActiveCfg = Debug|x64
+		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|Any CPU.Build.0 = Debug|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|x64.ActiveCfg = Debug|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|x64.Build.0 = Debug|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|x86.ActiveCfg = Debug|Win32
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Debug|x86.Build.0 = Debug|Win32
+		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Release|Any CPU.ActiveCfg = Release|x64
+		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Release|Any CPU.Build.0 = Release|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Release|x64.ActiveCfg = Release|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Release|x64.Build.0 = Release|x64
 		{C8FD1AA4-508D-4DC8-A239-F5B8DC4FF6B9}.Release|x86.ActiveCfg = Release|Win32
