+++ AnyRegCore/AnyRegCore.vcxproj
@@ -21,13 +21,13 @@
   <PropertyGroup Label="Globals">
     <VCProjectVersion>17.0</VCProjectVersion>
     <Keyword>Win32Proj</Keyword>
-    <ProjectGuid>{c8fd1aa4-508d-4dc8-a239-f5b8dc4ff6b9}</ProjectGuid>
-    <RootNamespace>AnyReg</RootNamespace>
+    <ProjectGuid>{7CEACDB6-D536-4CEA-AFE2-0D91A2203593}</ProjectGuid>
+    <RootNamespace>AnyRegCore</RootNamespace>
     <WindowsTargetPlatformVersion>10.0</WindowsTargetPlatformVersion>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
   <PropertyGroup Label="Configuration">
-    <ConfigurationType>Application</ConfigurationType>
+    <ConfigurationType>StaticLibrary</ConfigurationType>
   </PropertyGroup>
   <Import Project="$(SolutionDir)Configuration.props" />
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
@@ -41,7 +41,6 @@
   </ImportGroup>
   <PropertyGroup Label="UserMacros" />
   <ItemGroup>
-    <ClCompile Include="AnyReg.cpp" />
     <ClCompile Include="AnyRegDbConnection.cpp" />
     <ClCompile Include="Registry.cpp" />
     <ClCompile Include="WinError.cpp" />
