+++ AnyRegCore/Registry.cpp
