+++ AnyRegCore/Registry.hpp
