+++ AnyRegCore/WinError.cpp
