+++ AnyRegCore/WinError.hpp
